refactor(schedule): extract shared HabitChip pill from month-calendar
- New presentational HabitChip component (status glyph + client-owned
  badge + status classes) in schedule/habit-chip.tsx, rendering a button
  when onClick is given and a non-interactive span otherwise
- DayCell and ClientDayCell habit render sites now use HabitChip;
  class strings byte-identical, so the Agenda calendar is visually
  unchanged
- Removed now-unused habitChipClass/HabitStatusGlyph/ClientOwnedBadge
  and the User icon import (STATUS_PALETTE stays — shared with workouts)

Ref: https://github.com/mdb1/gc-fitness/issues/447
